fix useEndpoint setting order
